Moved some more classes around.
